requirements.txt
dj-database-url==0.5.0
Django==3.0.8
django-extensions==3.0.3
django-debug-toolbar==2.2
django-modeltranslation==0.15.1
django-photologue==3.11
django-sitetree==1.15.1
django-storages==1.9.1
django-taggit==1.3.0
django-tinymce==3.0.2
django-translation-manager==1.1.1
django-utils==0.0.2
ipython==7.16.1
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.5
raven==6.10.0
